requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.0
torchaudio==2.1.0
openai-whisper==20231117
SpeechRecognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
librosa==0.10.1
soundfile==0.12.1
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0